bot-threats-discord/requirements.txt
discord.py
feedparser
python-dotenv
